+++ source/DSCResources/DSC_WaitForDisk/DSC_WaitForDisk.psm1
@@ -37,7 +37,7 @@ function Get-TargetResource
         $DiskId,
 
         [Parameter()]
-        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName')]
+        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName','SerialNumber')]
         [System.String]
         $DiskIdType = 'Number',
 
@@ -94,7 +94,7 @@ function Set-TargetResource
         $DiskId,
 
         [Parameter()]
-        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName')]
+        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName','SerialNumber')]
         [System.String]
         $DiskIdType = 'Number',
 
@@ -176,7 +176,7 @@ function Test-TargetResource
         $DiskId,
 
         [Parameter()]
-        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName')]
+        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName','SerialNumber')]
         [System.String]
         $DiskIdType = 'Number',