+++ source/Modules/StorageDsc.Common/StorageDsc.Common.psm1
@@ -160,14 +160,14 @@ function Get-DiskByIdentifier
         $DiskId,
 
         [Parameter()]
-        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName')]
+        [ValidateSet('Number','UniqueId','Guid','Location','FriendlyName','SerialNumber')]
         [System.String]
         $DiskIdType = 'Number'
     )
 
     switch -regex ($DiskIdType)
     {
-        'Number|UniqueId|FriendlyName' # for filters supported by the Get-Disk CmdLet
+        'Number|UniqueId|FriendlyName|SerialNumber' # for filters supported by the Get-Disk CmdLet
         {
             $diskIdParameter = @{
                 $DiskIdType = $DiskId